+++ microros_transport/microros_transport.h
@@ -18,8 +18,9 @@ bool microros_transport_close(struct uxrCustomTransport *transport);
 size_t microros_transport_write(struct uxrCustomTransport *transport, const uint8_t *buf, size_t len, uint8_t *errcode);
 size_t microros_transport_read(struct uxrCustomTransport *transport, uint8_t *buf, size_t len, int timeout_ms, uint8_t *errcode);
 
-/* application API - call to publish string data to vision computer */
+#ifdef MICRO_ROS_ENABLED
 #include <rcl/types.h>
-rcl_ret_t microros_publish_string(const char *data);
+rcl_ret_t microros_publish_vision(void);
+#endif
 
 #endif